Adjust method for trimming ansible output debug information
This patch adjusts the method introduced in
I98fb57a56c1b7ddd6558e265cf6e65b159a05123 to trim the debug information
from the ansible output.

This adjustment allows the output to come to the console immediately,
instead of piping to a file and doing the processing after the ansible
run is complete. It also ensures that the return code passes through the
output trimming.

This commit combines the following cherry-picks:
- ed090eb1094472dcef755e6ca319310e5a0e7eac
- b76d43e8c60bcb3060ee4ee99b6fac8acf0650fb

Partial-bug: #1425482
Closes-bug: #1432295
Co-Authored-By: Hugh Saunders <hugh@wherenow.org>
Change-Id: I515adcad34ef76b8ea47fe90bc6572fb3aec1d2c