Adjust method for trimming ansible output debug information
This patch adjusts the method introduced in
I98fb57a56c1b7ddd6558e265cf6e65b159a05123 to trim the debug information
from the ansible output.

This adjustment allows the output to come to the console immediately,
instead of piping to a file and doing the processing after the ansible
run is complete.

Change-Id: I515adcad34ef76b8ea47fe90bc6572fb3aec1d2c
Partial-bug: #1425482