Simplify and improve bootstrap/gate/run scripts
Originally the gate check scripts were built in such a way that they
could cater for all branches with the same script. This no longer needs
to be the case as the gate checks are acknowledged to be branch-specific.

The necessary bits to cater for the other branches have now be removed
to simplify the scripts.

Also, in master all environment variables are exported, which helps with
ensuring that the variables are present when other scripts are also run.
This is a better practise and has been included in the script clean-up.

Change-Id: Ie0c05b1b94d167f296f0449531e76b16010fe1c0
Closes-Bug: #1431361