Adds the ability to define repositories within a container
This change implements the ability to set the apt repositories
within a container upon container create from both a deployed
image as well as when using the lxc-openstack container build
template.

* There are two new namespaced variables that set the main
  and security repositories for Ubuntu. These two variables
  are the same in both the `lxc_container_create` and `lxc_hosts`
  roles. These two new variables have been set as defaults in their
  respective roles.

* The method by which the template and container create deploy the
  repositories within the container is the same. The old sources list
  is removed and the new entries are created. This will provide a
  consistent experience between hosts and containers as they can now
  use the same repositories without human intervention.

With this change it will now be possible to deploy containers within
a walled garden where the apt repository may not be publicly
accessible.

Juno-Backport-Potential
Closes-Bug: #1433412
Change-Id: Iaf51a5e3919f684b20da370a9470dc45e2452d4e