Remove logstash multiline filter and change default number of workers
This patch removes the use of the multiline filter in the logstash
configuration and changes the default number of workers used to the
number of CPU's reported by ansible. The number of workers can still be
set manually if so desired.

This is specifically to improve logstash event processing speed which is
far too slow for a multi-server environment to be useful in any way.

Change-Id: Id9fbab6b0f6e513c7dcae4e2713555840f9d996a
Closes-Bug: #1423755
(cherry picked from 9cb3b4184fcd4edd6b895914425ed403fd76eeac)