Remove support for ~/.openstack_deploy/
The README documents support for housing openstack_deploy/ in an
unprivildged user's home directory, but then the script for creating
openstack-ansible immediately requires root privileges, meaning that you
can't setup openstack_deploy/ in an unpriviledged user's home dir at
all.

This removes support for the feature instead of trying to fix it,
because I don't see a use case in keeping it, and there are many
additional fixed references to /etc/keystone_deploy/ that would have to
be made dynamic.

Change-Id: I19c6a568d1d52c33b3c13e9c007ca59e44d8c1dd
Closes-Bug: 1435597