Fix dedicated replication network logic
When the replication network is the same as the storage network the
conf should not be dropped and the appropriate configuration stanzas
should exist.
Additionally, the replicator init script needs to be placed, and when
using dedicated replication network the replicator-server conf needs to
be populated appropriately.
Adds a further check to make it obvious when the dedicated conf is not
being used and clean up the conf if it exists.

Closes-Bug: 1435802
Change-Id: I3befc1b24f3625b0cfd815afdfab981101d2fd8e