Add support for passing api version to glance module
Glance API version 1 is deprecated as of kilo.  If we want to support
v2 moving forward, we should add support for passing a version in to
the glance module, and alter the module to take appropriate action
depending whether we are using v1 or v2.  This commit does that, defaulting to 1 to keep
compatibility with current code in master.

Change-Id: Iac9dcd3f5fdf1cfacde3e9cb10cb1aa2b72500a8
Closes-Bug: 1435808