Move SYMLINK_DIR declaration to bootstrap-aio.sh
The gate-check-commit.sh script defined the SYMLINK_DIR variable, but it
was only used in bootstrap-aio.sh. This meant that using
bootstrap-aio.sh by itself would fail due to the undefined variable.

This change moves the declaration to boostrap-aio.sh so that an AIO can
be boostrapped without necessitating a gate check.

Change-Id: Id1b20379c5837cdd4b3df03de42f0ea612e3fbb5
Closes-Bug: #1434315