Update rabbitmq_package_url
Changed the rabbitmq package url to point to the pinned version of rabbitmq
as found within the https://mirror.rackspace.com/rackspaceprivatecloud repo.
This change is a stabilization change that will ensure that we have a
mirrored copy of the version of rabbitmq that we believe is production ready.

Change-Id: I55b28f073a3ed3e0924cdd0007149bb1d9de21f7
(cherry picked from commit a7d245ec4d917cb2659157df8c56edddaf911d43)
Closes-Bug: #1436240