Fix bug in playbooks/library/neutron
This change updates _add_router_interface to list ports for the router
device only, and not for all devices.  As we select specific device
ports only, we no longer need to skip over other device's ports, and
this indirectly fixes a bug where we were using 'next' instead of
'continue' to skip to the next iteration of the for loop.

Change-Id: I892dc0f58f1d8bbb7ebd46f7790c5420c4c8b94b
Closes-Bug: #1439295