Add network.yml monitoring playbook
To monitor the number of open connections a new monitoring playbook has
been added called network.yml. This playbook will need to be run as
part of the deployment when the other monitoring playbooks are used.

This playbook will create a new alarm that will alert when > 90% of
nf_conntrack_max is reached by nf_conntrack_count. The data is
collected using the conntrack_count.py plugin in the rcbops-maas repo.

Closes-Bug: #1439604
Change-Id: I8aed6df1157f5337b531ba24ec566200b6e67af3