Adjust AIO swap size to 4G/8G, depending on RAM amount
The AIO setup used for the commit checks in the gate is highly resource
constrained. Currently the swap space available is 1GB and the RAM available
is 8GB. The swap size would be better set to 4G if the RAM amount is 8GB or
less, or 8G if the RAM amount is more than 8GB.

The patch switches from using dd to using fallocate for the thick loopback
file creation in order to prevent memory exhaustion when trying to create
files that are larger than the available memory.

Change-Id: I9cdb145780fb867b933890111f642a48ed7add70
Closes-Bug: #1442028