Remove md5sum check for environment vs user_config
The md5sum check doesn't currently solve any issues we have, and with
the addition of a custom "env.d" directory this will only cause
frustration as the md5sum changes.

The initial idea was to prompt a user to update their
user_config/user_variables/environment files between updates in the
os-a-d repository but the md5sum check never really solved this problem,
since both the files being checked won't have changed if you update the
repo and don't copy the new versions over.

Change-Id: Id1e7d307aa5dbffe069d6d4fa4569dd13ad4e8c1
Partial-Bug: #1399430