Added script to facilitate juno to kilo upgrade
This script was created to allow for Juno (v10.x) to be
upgrade to Kilo (v11.x). This script attempts to wrap all
of the changes that have taken place between the Rackspace
release of Juno and the community product created in master
that is Kilo.

Partially implements: blueprint master-kilofication

Change-Id: Ibbc2a8192fb3a4e8508aefc74b1d062cd890f1a0