Add release pinning to galera roles
As both Ubuntu and the MariaDB repositories carry packages with the same
names, apt package pinning is required to ensure that the packages we install
come from the right repository. This patch adds a pinning preference file to
ensure that the packages we install come from the MariaDB repository.

Due to the OpenStack CI images having conflicting packages installed in the
base image, this patch also includes the removal of these packages during the
AIO bootstrap to ensure that the deployment succeeds.

For environments where the same two packages are installed on the hosts prior
to the deployment of os-ansible-deployment, they would need to ensure that the
packages are removed first. This has been seen in environments where
monitoring tooling has installed the packages as a dependancy.

Closes-Bug: #1446772
Change-Id: I5ad0b9010e49832c4f866c0b141e1c035d5f9268