Remove known conflicting packages in the AIO bootstrap
Due to the OpenStack CI images having conflicting packages installed in the
base image, this patch also includes the removal of these packages during the
AIO bootstrap to ensure that the deployment succeeds.

For environments where the same two packages are installed on the hosts prior
to the deployment of os-ansible-deployment, they would need to ensure that the
packages are removed first. This has been seen in environments where
monitoring tooling has installed the packages as a dependancy.

Closes-Bug: #1446772
Change-Id: I83dda553dec9b7652d8463c426b2de6724ba78f2
(cherry picked from commit 66cdfc555c16ed0887ba04ed6e91334da20060da)