Add newline to /etc/ssh/sshd_config if necessary
On some deploys we see /etc/ssh/sshd_config missing a trailing newline,
which breaks the sshd_config file when the lineinfile tasks are run.
This has been traced back to the version of ansible being deployed in
juno as we do not have this issue in later branches using newer
versions of ansible.

Alternatively, we could upgrade ansible in juno but I would prefer to
make the smallest change possible since juno is now widely deployed.

Closes-Bug: #1442239

Change-Id: If0f66753f40b378f6f2c8d256e1eb5b4d59fd64e