Removed hard coded package pinning for mariadb/galera
This change removes the hard set package pinning from the upstream
repository. The hard coded nature of this makes it impossible to
maintain over the lifecycle of a given release. While the default
pinning has been removed the ability to pin to whatever the
deployer wants still exists and can be used as the deployer sees
fit.

Change-Id: I94300184f28b9564dedcc7d14f18aa793765ed8f
Closes-Bug: #1446772