Add ubuntu backports repository to AIO if it's not present
The swift installation process requires access to the Ubuntu backports
repository to install the liberasurecode1 and liberasurecode-dev packages. In
the hpcloud environments used in OpenStack-CI the backport repository is not
implemented, so we need to ensure that it's there if it hasn't already been
configured. This patch does that as part of the AIO bootstrap process.

Additionally, heat now requires python-saharaclient >= 0.9.0 and we bump here
to allow heat to install successfully.

Note that both issues are causing the gate to fail and need to be addressed
together to get a successful gate.

Closes-Bug: #1448126
Closes-Bug: #1448152
Change-Id: I5d8c6f4eb7182129ab22defd70a50d78e7ffe86f