Updated juno for new dev work
Updated release revision in Juno for all services.
This changes the requirement.txt to match what is in upstream
OpenStack for the various services and adjusts pins / removes
items that are no longer needed / relevant.

The elasticsearch github organization has renamed itself to
"elastic" and this change reflects the new organization name
in the repo package files.

Change-Id: I87091eba96859e9a1b164776ef52034b7a70405d