Updated host repositories when using the AIO
This change forces the apt mirrors to use a set that is known good.
The previous changes to set the mirrors to the same as what was
already on the host has proven itself to be a greater issue by not
using a set that is known to be in good working order.

Change-Id: I1872a32b53b9caecb8ac0803db304fa0ee254475
Closes-Bug: 1441148
(cherry pick from commit 1a5562d04a62c318ba6961b114cefadb748f1a9b)