Harden our copy of Glance's policy
Most of Glance's current checks are implemented in the API controllers
but in Kilo, Glance added the ability to actually define meaningful
policy rules around images and image members. In an effort to harden
our default config as best as we can, we should check to see if the
user trying to perform some of these actions are either an admin or the
owner of the image.

Change-Id: I2dcf4d828c9be88143174de30a6b59d655ab0539
Closes-bug: 1408363
(cherry picked from commit 8bebbc6e530efe26c98f396b24c0ea00971093d3)