Add Sphinx machinery and initial docs.
This commit gets the doc directory and Sphinx configuration started.

Further documentation will depend on the Sphinx files and the index.rst
file at the very least.

The dev-requirements were updated to install Sphinx, and a gating script
was added to do the building for when we move to actually do gating on
the docs. The docs gate script is currently set to treat warning as
errors.

Change-Id: I977a2e53f1b0981cbc702ac53f2198d9f6a9da9d
Implements: blueprint developer-docs