Unblock master gating
Most of master is transitioning to the liberty cycle, to allow us
to continue gating the service install versions is being set back
to the head of the stable/kilo branch. This should be updated once
all upstream OpenStack projects fully transition to Liberty.

This adds a wait condition when creating the containers to ensure
that the containers are created correctly and that ssh is available
within them before actually continuing with the installation process

This change forces the apt mirrors to use a set that is known good.
The previous changes to set the mirrors to the same as what was
already on the host has proven itself to be a greater issue by not
using a set that is known to be in good working order.

Change-Id: I1872a32b53b9caecb8ac0803db304fa0ee254475
Closes-Bug: 1441148