Lock down PBR to < version 0.11.0
This change was made to ensure that PBR is using version < 0.11.0.
This is because changes in upstream PBR have made it impossible to
build contributing packages in heat and will likely impact other
services until the switch to liberty is in full effect. To ensure
consistent building of OpenStack packages PBR is now being installed
in the repo server containers which will ensure there are no
assumptions being made when creating python packages from OpenStack
contributing sources that may not be as well maintained as others.

Change-Id: I20c216543e63ef142119c2625fb219e030a708dd
Partial-bug: 1450733