Fixed container user create issue
This patch resolves an issue where the LXC container create user
password is being hard coded due to how the default LXC container
template is being used. The template allows for a container to be
created without specifying a password at create time however when
doing this the default LXC template will create a user and a
password based on the linux distribution's name. This means that
all users of a "ubuntu" template will have a "ubuntu" user which
will have a "ubuntu" password. While the container user must exist
the password does not need to be known to the end user/deployer
because there are other means to gain access to a container from
within the host. So to correct this issue for all deployers as
well as all future deployments this patch ensures that the container
user is created with a random password and provides for an additional
tag to allow the deployer to regenerate the password at will.

Change-Id: I46d8d407028485ad99f89347bf93354ed5043150
Closes-Bug: 1437054