Updated Juno for new dev cycle
The base change here is that Juno can now being using newer setuptools
and pip6. This is because the logical issues that prevented juno from
being able to upgrade in the past have been fully resolved upstream.

Glance store has also been pinned because there is no stable/juno
branch and we do not want the service to be upgraded oustide the
confines of what should would within Juno.

rackspace_monitoring_cli was updated because it had a packaging
problem in v6.0.6 that prevented from being built properly which was
overlooked in previous version of pip.

oslo.log needs to be updated to support tempest becuase pip6 is
for real resolving its deps where is wasn't so much in pip 1.5.6.

Closes-bug: 1452305
Change-Id: Ieeede35c8acf46b2ef27fa1e0e17944307c2d97f