Set all groups when creating system_user
Previously we had a task in nova_libvirt to add the nova user to the
libvirtd group. The problem with this is that if the user is recreated,
it will be removed from the group, as that group is not specified when
the user is created.

This patch creates and sets additional groups when the user is created
so any recreation will be idempotent.

Change-Id: I62eedca6d417fbf336538452c9869de5b58fb040
Closes-Bug: 1442366