Move Galera logs to /var/log/mysql_logs and set correct ownership
By default the permissions on the Galera log files are too restrictive
for rsyslog to read.  This patch moves those files to /var/log/mysql_log
and gives it group ownership of syslog.  Further it sets the sticky bit
on /var/log/mysql_log and ensures that the syslog group has read
access.

Change-Id: I57403a4147a07f949c78e648357070e23fe6d47f
Closes-bug: 1452316