Update for pip 7.0.0
The trusted host option when installing from a domain that is not
https is no longer optional. This change adds a domain variable to
the groupvars/all.yml and ensures that the get-pip.py installation
process is using the `--trusted-host`` flag and is pointed back at
the domain name used in the "openstack_upstream_domain".

Change-Id: I18539cf4b8790c096d44a0de4985b9d4293d95e3