Avoid keeping re-downloading container caches
When the cache tarball gets deleted at the end the task set in
playbooks/roles/lxc_hosts/tasks/lxc_cache.yml
then it will be downloaded again at subsequent runs, which makes the
download task not idempotent, and needlessly slow.

The unarchive task also needs to only be executed if the download task
performed a change.

Ensure that the location the download is placed and the directory it's extracted
to are different.

Change-Id: I4e433e82499f5cfa0b5dde44a6aa7bab113d9e0e