Restart Apache when Keystone configuration changes
Currently, when a change occurs in a Keystone configuration file it's not
taking effect since the Apache service is not being notified to restart.

Merge the keystone_apache role with the keystone_common role to take
advantage of the existing Restart Apache handler with tasks that
make changes to Keystone configuration files.

Change-Id: If7bbe23688bbc0565e43a74103a9d9ce89a82236
Closes-Bug: #1457737