Restart Apache when Keystone configuration changes
Currently, when a change occurs in a Keystone configuration file it's
not taking effect since the Apache service is not being notified to
restart.

Merge the keystone_apache role with the keystone_common role to take
advantage of the existing Restart Apache handler with tasks that
make changes to Keystone configuration files.

Also remove unused roles/keystone/files.

Change-Id: I6c5e920741d66b3ead356f372dfbd68461e3f854
Closes-Bug: #1457737