Add alarm for rabbitmq_max_channels_per_conn
This commit adds a new alarm for metric rabbitmq_max_channels_per_conn,
which is a metric representing the largest (max) number of channels per
connection across all rabbitmq connections. We have chosen a threshold
of 10 -- running a dev cluster periodically sees
rabbitmq_max_channels_per_conn > 1 and 10 is an 'arbitrary but likely
reasonable limit' which we can use to prevent channel ballooning.

Change-Id: I40bd5dde25d61e2febb8699dbcd1e28b0f25674c
Closes-Bug: #1459249