Backported some bootstrap processes for gating
This change backports the ansible install process from master when
using the `bootstrap-ansible.sh` script. This was backported
because the pip install process for ansible version 1.6.10 no longer
works due to upstream changes in the location of the ansible core
modules.

This also adds logging to the gate run and cleans up the console
output so that we can better understand what is going on.

This change should resolve the issues where the ansible core modules
are returning an error stating the following:
 "is not a legal parameter in an Ansible task or handler"

Change-Id: I03160f8159b465f9cb1381312cdc65e572fb2dbd