Updated repo server deps
This updates the deps that need to be on the repo server to ensure
that all of the python sources are built correctly. The additional
packages are in response to an upstream package dep
`python-qpid-proton` which is now requiring these items.

Change-Id: I6d6f70734fd46c4055f49fdf25035d2b4b212414
(cherry picked from commit 125b0d01b0dc0274fad120f6be0844bddccd0b89)