Moved all of group_vars into a variable file
Moved all of the group_vars/all.yml file into its own variable file
This change was done to allow a user to override basic options
without having to modify the default group variable files. While
the group_vars/all.yml file is still present it is only holding the
revision information that is used for release data and the minimal
required kernel that allows the system to function using VXLAN.

The upgrade script was modified to support the new "default"
user_group_vars.yml file.

tempest_swift_enabled was set to true in group_vars, so this has
now been set as a default in the role instead.

Commit 1bd2bc052a36c7f8c28e473d0a37fd29827198b0 implemented the
package URL update for rabbitmq, but not the corresponding sha256
update. This was not noticed due to group_vars overriding the URL
to a previous version, resulting in the above-mentioned commit
taking no effect. This patch therefore also corrects the sha256.

Conflicts:
        playbooks/inventory/group_vars/all.yml

Closes-Bug: #1460516
Closes-Bug: #1460992
Change-Id: I8e42bb124827bb276134d662c9a171db8e4c017e
(cherry picked from commit eb95036a7ea0e98fc931dd19f416015c66f3620c)