Updated upgrade and deployment processes for https
This change simply makes sure that all containers and hosts have the
`apt-transport-https` package installed. this package is absolutly
required everywhere because we've changed all of the repo endpoints
to https and not all systems have this package installed on a base
kick. Furthermore an entry was added into the upgrade script to
ensure that upon upgrade everything will converge and remain
consistent.

Change-Id: I4b357ff7099a4c1c63c85ac9560aefc8d56709be
Closes-Bug: #1463155