Implement service_region user_group_var
Currently each service has an individual service_region var
which needs to be set when setting up a region that is
different to the default.

This patch implements a user_group_var called 'service_region'
which can be used to set the region for all services.

The AIO bootstrap script is also altered to set the
service_region in the user config.

Change-Id: I5ed216916f667506ffed900a01744dad973221bf
Closes-Bug: #1463366