Allow Horizon to access multiple regions
This patch introduces the capability for Horizon to access
multiple regions.

This only takes effect if horizon_available_regions is defined,
otherwise the current default comment is applied to the
local_settings file.

Usage is described via comments in the role defaults and in the
etc/openstack_deploy/user_variables.yml file.

Change-Id: Ia8c0b8be67fd0f479c492d9919b542806f9b204f