Create Junit XML Report from tempest run
Tempest results are currently written to stdout as part of a test run.
This output is difficult to parse. This patch adds steps to generate an
XML report, and make this report available in /tmp on a physical host.

The report is copied to a physical host so that external systems that do
not have knowledge of the container inventory can retrieve the report,
for example Jenkins.

* Add pip packages python-subunit and junitxml to tempest venv, needed
  to convert the subunit stream stored in testrepository to junit XML.

* openstack_tempest_gate:
    After a tempest run completes, the subunit stream is extracted from
    the testrepository, converted to junit XML and written to /tmp in
    the utility container running the tests.

* run-tempest.sh
    After running openstack_tempest_gate, retrieve XML results and write
    them to /tmp on the host.

Change-Id: I897024f545dd2a90f80b0d5c0aa558f335dc963a
Closes-Bug: #1464686
(cherry picked from commit 8a7d5179d66cd43c3f49cec6a764f1073232e6f3)